[FEATURE] #26 - 파라미터 orderId의 타입을 Long으로 수정

+++ app/src/main/java/co/kr/woowahan_banchan/data/database/dao/OrderItemDao.kt
@@ -9,10 +9,10 @@ import kotlinx.coroutines.flow.Flow
 @Dao
 interface OrderItemDao {
     @Query("SELECT * FROM ORDER_ITEM WHERE order_id LIKE (:orderId)")
-    suspend fun getItems(orderId: Int): List<OrderItemDto>
+    suspend fun getItems(orderId: Long): List<OrderItemDto>
 
     @Query("SELECT COUNT(id) FROM ORDER_ITEM WHERE order_id LIKE (:orderId)")
-    suspend fun getItemCount(orderId: Int): Int
+    suspend fun getItemCount(orderId: Long): Int
 
     @Insert
     suspend fun insertItems(items: List<OrderItemDto>)
